+++ src/main/java/techcourse/fp/chess/MainApp.java
@@ -2,8 +2,9 @@ package techcourse.fp.chess;
 
 import java.util.Scanner;
 import techcourse.fp.chess.controller.ChessController;
-import techcourse.fp.chess.dao.LocalMysqlChessGameDao;
+import techcourse.fp.chess.dao.MysqlChessGameDao;
 import techcourse.fp.chess.dao.MysqlPieceDao;
+import techcourse.fp.chess.service.ChessGameService;
 import techcourse.fp.chess.view.InputView;
 import techcourse.fp.chess.view.OutputView;
 
@@ -14,7 +15,7 @@ public final class MainApp {
     public static void main(String[] args) {
         try (Scanner scanner = new Scanner(System.in)) {
             final ChessController chessController = new ChessController(new InputView(scanner), new OutputView(),
-                    new ChessGameService(new LocalMysqlChessGameDao(), new MysqlPieceDao()));
+                    new ChessGameService(new MysqlChessGameDao(), new MysqlPieceDao()));
             chessController.run();
         } catch (Exception exception) {
             exception.printStackTrace();
