+++ src/main/java/techcourse/fp/chess/service/ChessGameService.java
@@ -1,4 +1,4 @@
-package techcourse.fp.chess;
+package techcourse.fp.chess.service;
 
 import java.util.List;
 import java.util.Map;
